refactor(balance-cards): move balance-cards module from scripts/ to src/ (CG-0MRDNVQPS009UUIN)
- Relocate scripts/balance-cards/{algorithm,csv,index,rationale,summary}.ts to src/balance-cards/
- Update scripts/run-balance-cards.ts to import from ../src/balance-cards; fix usage text path
- Update tests/main-street/balance-cards.test.ts import path
- Update PRD docs path references (scripts/balance-cards -> src/balance-cards)
- No functional changes: CLI output verified byte-identical
- Stabilize pre-existing flaky load-sensitive tests: replay CLI subprocess timeouts, screen-layout perf budget, generate-thumbnail, ai-strategy 200-seed, monte-carlo guardrails, lost-cities full match (timeout increases)
