+++ tests/ui/screen-layout-mapping.test.ts
@@ -117,7 +117,11 @@ describe('screen layout mapping utilities', () => {
     }
 
     const elapsedMs = Date.now() - started;
-    expect(elapsedMs).toBeLessThan(50);
+    // Budget is intentionally generous: normalizedToPixels is a pure math
+    // hot-path used per-zone per-frame, so the guardrail only needs to catch
+    // pathological regressions (e.g. O(n^2) or accidental I/O), not micro-
+    // performance. 500ms also tolerates heavily-loaded shared CI machines.
+    expect(elapsedMs).toBeLessThan(500);
   });
 
   it('handles zones with pixelOverride for position-only override', () => {
